+++ R/PrepData.R
@@ -6,7 +6,8 @@ PrepData <- function(data, model, itemtype, guess, upper, gpcm_mats,
     if(is.null(grsm.block)) grsm.block <- rep(1, ncol(data))
     # if(is.null(rsm.block)) rsm.block <- rep(1, ncol(data))
     itemnames <- colnames(data)
-    keywords <- c('COV', 'CONSTRAIN', 'CONSTRAINB', 'PRIOR', 'MEAN', 'START', 'LBOUND', 'UBOUND')
+    keywords <- c('COV', 'CONSTRAIN', 'CONSTRAINB', 'PRIOR', 'MEAN', 'START', 'LBOUND', 'UBOUND',
+                  'FIXED')
     data <- as.matrix(data)
     colnames(data) <- itemnames
     J <- ncol(data)